+++ src/react/index.ts
@@ -19,6 +19,7 @@ export { useIdle } from './use-idle';
 export { useWhenIdle } from './use-when-idle';
 export { useCanvas } from './use-canvas';
 export { useMutation } from './use-mutation';
+export { usePointer } from './use-pointer';
 export { useTween } from './use-tween';
 export { usePresence } from './use-presence';
 export { Presence } from './presence';
@@ -71,12 +72,14 @@ export type {
 } from './use-canvas';
 export type {
   MutationRecordsCallback,
-  MutationPhaseCallback,
   UseMutationOptions,
-  UseMutationReactiveResult,
-  UseMutationTransientResult,
   UseMutationResult,
 } from './use-mutation';
+export type {
+  PointerCallback,
+  UsePointerOptions,
+  UsePointerResult,
+} from './use-pointer';
 export type { UseTweenOptions } from './use-tween';
 export type {
   PresenceMode,